Use python-osc module and log with logging module
